+++ SpawnDev.SpawnJS/Marshallers/StringMarshaller.cs
@@ -6,7 +6,7 @@ namespace SpawnDev.SpawnJS.Marshallers
     public class StringMarshaller : JSMarshallerFromString<string>
     {
         public override string JSToNet(string value) => value;
-        public override void NetToJS(SpawnJSObjectReference jsParent, double jsKey, string value) => jsParent.PropertySet(jsKey, value);
+        public override void NetToJS(SpawnJSObjectReference jsParent, int jsKey, string value) => jsParent.PropertySet(jsKey, value);
         public override void NetToJS(SpawnJSObjectReference jsParent, string jsKey, string value) => jsParent.PropertySet(jsKey, value);
     }
 }
